refactor(pine): move the lowering passes out of the compile driver
`compile.rs` was 1,540 production lines carrying all six passes behind
banner comments. Each banner is now a module, by pure move, so the file
that says what the passes are is not also the file that runs them.

`compile/declarations.rs` owns pass 1 — the `indicator()` header, the
top-level `plot*` registry, `fill()` resolution and the plot defaults
and `style_of_tag` only it reads. `compile/fold.rs` owns pass 4's
folder: `Const`, the `color`/`plot`/`shape`/`location` member constants
and `Compiler::fold`, which both other passes call.
`compile/resolve.rs` owns passes 2, 3, 5 and 6 — the one resolved walk,
with `Scope`, `WalkCtx`, `Frame` and `source_of_builtin`, which only
`input_call` reads. `compile.rs` keeps the public surface
(`CompiledScript`, `Resolution`, `FunctionInfo`, `CallSiteId`,
`compile()`) and the `Compiler` struct the passes run over.

Every body moved unchanged; the only widenings are `pub(super)` on the
eleven items a sibling or the driver names across the seam.
`compile.rs` falls to 323 production lines; the largest sibling is 689.
